Migrate to the standalone ApodiniTypeInformation framework
