+++ .gitignore
@@ -1,5 +1,7 @@
 # Apodini .gitignore File
 
+.idea
+
 # Swift Package Manager
 Package.resolved
 *.xcodeproj
