+++ Sources/ApodiniMigratorCore/TypeInformation/TypeInformation+Value.swift
@@ -1,0 +1,11 @@
+//
+//  TypeInformation+Value.swift
+//  ApodiniMigratorCore
+//
+//  Created by Andreas Bauer on 23.08.21.
+//  Copyright © 2021 TUM LS1. All rights reserved.
+//
+
+import ApodiniTypeInformation
+
+extension TypeInformation: Value {}
