+++ Sources/ApodiniMigratorShared/Extensions/Array+Extensions.swift
@@ -36,19 +36,6 @@ public extension Sequence {
 }
 
 public extension Array where Element: Equatable {
-    /// Returns whether self is equal to other not considering the order of elements
-    func equalsIgnoringOrder(to other: Self) -> Bool {
-        guard count == other.count else {
-            return false
-        }
-
-        for element in self where !other.contains(element) {
-            return false
-        }
-
-        return true
-    }
-    
     /// Returns a new array in which all occurrences of a `target` element are replaced with `replacement` element.
     func replacingOccurrences(ofElement target: Element, with replacement: Element) -> Self {
         reduce(into: Self()) { result, current in
