+++ Sources/ApodiniMigratorShared/Extensions/Encodable+Extensions.swift
@@ -48,13 +48,3 @@ public extension Encodable {
         return filePath.absolute().string
     }
 }
-
-// MARK: - KeyedEncodingContainerProtocol
-extension KeyedEncodingContainerProtocol {
-    /// Only encodes the value if the collection is not empty
-    public mutating func encodeIfNotEmpty<T: Encodable>(_ value: T, forKey key: Key) throws where T: Collection, T.Element: Encodable {
-        if value.isNotEmpty {
-            try encode(value, forKey: key)
-        }
-    }
-}
