+++ Tests/ApodiniMigratorTests/ApodiniMigratorCore/ApodiniMigratorModelsTests.swift
@@ -3,25 +3,6 @@ import XCTest
 @testable import ApodiniMigratorClientSupport
 
 final class ApodiniMigratorModelsTests: ApodiniMigratorXCTestCase {
-    func testFluentProperties() throws {
-        try FluentPropertyType.allCases.forEach { property in
-            let json = property.json
-            let instance = XCTAssertNoThrowWithResult(try FluentPropertyType.decode(from: json))
-            XCTAssertEqual(property.description, property.debugDescription)
-            XCTAssert(instance == property)
-        }
-        
-        XCTAssertThrows(try FluentPropertyType.decode(from: "@FluentID".json))
-        
-        let childredMangledName = MangledName("ChildrenProperty")
-        if case let .fluentPropertyType(type) = childredMangledName {
-            XCTAssert(type == .childrenProperty)
-            XCTAssert(type.isGetOnly)
-        } else {
-            XCTFail("Mangled name did not correspond to a fluent property")
-        }
-    }
-    
     func testDSLEndpointIdentifier() {
         var errors: [ErrorCode] = []
         errors.addError(404, message: "Not found")
@@ -95,36 +76,7 @@ final class ApodiniMigratorModelsTests: ApodiniMigratorXCTestCase {
         
         XCTAssertThrows(try Version.decode(from: "v123".json))
     }
-    
-    func testRuntime() throws {
-        typealias User = TestTypes.User
-        let info = try info(of: User.self)
-        
-        let name = try RuntimeProperty(XCTAssertNoThrowWithResult(try info.property(named: "name")))
-        XCTAssert(name.ownerType == User.self)
-        XCTAssert(name.genericTypes.count == 1)
-        XCTAssert(!name.isIDProperty)
-        XCTAssert(!name.isFluentProperty)
-        
-        XCTAssert(info.cardinality == .exactlyOne(User.self))
-        
-        XCTAssert(type(User.self, isAnyOf: .struct))
-        XCTAssert(!type(type(of: ()), isAnyOf: .struct, .enum))
-        
-        let url = try XCTUnwrap(try createInstance(of: URL.self) as? URL)
-        XCTAssert(url == .default)
-        
-        enum TestError: Error {
-            case test
-            case opaque
-        }
-        
-        XCTAssert(!knownRuntimeError(TestError.test))
-        
-        XCTAssertThrows(try instance(User.self)) // can't create an enum instance
-        XCTAssertNoThrow(try instance(TestTypes.Shop.self))
-    }
-    
+
     func testDocument() throws {
         var document = Document()
         document.add(
@@ -147,12 +99,5 @@ final class ApodiniMigratorModelsTests: ApodiniMigratorXCTestCase {
         XCTAssertEqual(document.metaData.versionedServerPath, "http://127.0.0.1:8080/test1")
         XCTAssert(!document.json.isEmpty)
         XCTAssert(!document.yaml.isEmpty)
-    }
-    
-    func testNull() throws {
-        let null = Null()
-        let data = XCTAssertNoThrowWithResult(try JSONEncoder().encode(null))
-        XCTAssertNoThrow(try JSONDecoder().decode(Null.self, from: data))
-        XCTAssertThrows(try JSONDecoder().decode(Null.self, from: .init()))
     }
 }
