+++ Tests/ApodiniMigratorTests/ApodiniMigratorGeneratorTests/AuxiliaryFileGeneratorTests.swift
@@ -28,8 +28,8 @@ final class AuxiliaryFileGeneratorTests: ApodiniMigratorXCTestCase {
             name: .init(name: "TestEnumeration"),
             rawValueType: .string,
             cases: [
-                .case("first"),
-                .case("second")
+                .init("first"),
+                .init("second")
             ]
         )
